Completed Common API Refactoring
